Huge strides in dividing up the content into traditional divisions. Also made some good progress with the user interface side of it providing styles and layouts for the different views and widgets.
Provided an additional VerseBrowser view which focuses on just one verse
at a time.

Gesture controls added to allow quick navigation in the ScriptBrowser
and in the VerseBrowser.

Probably did too much for one commit but this may be close to ready for
initial release of the Manna Canon app.